requirements.txt
customtkinter==5.2.2
Pillow==9.5.0
requests==2.31.0
python-dotenv==1.0.1
mysql-connector-python==8.3.0
matplotlib==3.8.3
pandas==2.2.1 